+++ lib/elasticsearch_ex.ex
@@ -33,4 +33,11 @@ defmodule ElasticsearchEx do
   @spec index(source(), index(), nil | document_id(), keyword()) :: response()
   defdelegate index(source, index, document_id \\ nil, opts \\ []),
     to: ElasticsearchEx.Api.Document
+
+  @doc """
+  Refer to `ElasticsearchEx.Api.Search.search/2` documentation.
+  """
+  @doc since: "1.3.0"
+  @spec stream(query(), nil | index(), keyword()) :: Enumerable.t()
+  defdelegate stream(query, index \\ nil, opts \\ []), to: ElasticsearchEx.Stream
 end