+++ lib/elasticsearch_ex/stream.ex
@@ -13,11 +13,11 @@ defmodule ElasticsearchEx.Stream do
 
   @type index :: ElasticsearchEx.index()
 
-  @type pit :: %{required(:id) => binary(), required(:keep_alive) => binary()}
+  @typep pit :: %{required(:id) => binary(), required(:keep_alive) => binary()}
 
-  @type search_after :: []
+  @typep search_after :: []
 
-  @type acc :: {pit(), nil | search_after()}
+  @typep acc :: {pit(), nil | search_after()}
 
   ## Module attributes
 
@@ -30,9 +30,15 @@ defmodule ElasticsearchEx.Stream do
 
   ## Examples
 
-      iex> ElasticsearchEx.Stream.stream(%{query: %{match_all: %{}}, sort: [%{"@timestamp" => "desc"}]}, "my_index", keep_alive: "30s", per_page: 500)
+      iex> ElasticsearchEx.Stream.stream(
+      ...>   %{query: %{match_all: %{}}, sort: [%{message: :desc}]},
+      ...>   :my_index,
+      ...>   keep_alive: "30s",
+      ...>   per_page: 500
+      ...> )
       #Function<52.124013645/2 in Stream.resource/3>
   """
+  @doc since: "1.3.0"
   @spec stream(query(), nil | index(), keyword()) :: Enumerable.t()
   def stream(query, index \\ nil, params \\ []) do
     {keep_alive, params} = Keyword.pop(params, :keep_alive, "10s")