+++ lib/elasticsearch_ex/error.ex
@@ -19,13 +19,13 @@ defmodule ElasticsearchEx.Error do
   ## Public functions
 
   @impl true
-  @spec exception(AnyHttp.Response.t()) :: t()
-  def exception(%AnyHttp.Response{status: status, body: nil}) do
+  @spec exception(Req.Response.t()) :: t()
+  def exception(%Req.Response{status: status, body: nil}) do
     %__MODULE__{status: status, reason: "Response returned #{status} status code"}
   end
 
   @impl true
-  def exception(%AnyHttp.Response{status: status, body: %{"error" => error}}) do
+  def exception(%Req.Response{status: status, body: %{"error" => error}}) do
     %__MODULE__{
       status: status,
       reason: error["reason"],
@@ -36,7 +36,7 @@ defmodule ElasticsearchEx.Error do
   end
 
   @impl true
-  def exception(%AnyHttp.Response{
+  def exception(%Req.Response{
         status: 404,
         body: %{"_id" => doc_id, "result" => "not_found"} = body
       }) do